Target Consoles now have events to fire, mutex protection for debugger, Revamp events so they dont need to be re registered & Impliment all events on host side.

+++ Playstation/OrbisLibAPI/APIPackets.h
@@ -191,9 +191,9 @@ struct AppListPacket
 
 struct LibraryPacket
 {
-	int64_t Handle;
+	uint32_t Handle;
 	char Path[256];
-	int32_t SegmentCount;
+	int SegmentCount;
 	OrbisKernelModuleSegmentInfo Segments[4];
 };
